+++ components/automate-cli/cmd/chef-automate/ha_node_add_onprem_test.go
@@ -8,22 +8,77 @@ import (
 	"github.com/chef/automate/lib/majorupgrade_utils"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
+	"golang.org/x/crypto/ssh"
 )
 
-func PullConfFunc(sshUtil *SSHUtil, ex []string) (*ExistingInfraConfigToml, error) {
-	cfg, err := readConfig(CONFIG_TOML_PATH + "/config.toml")
-	if err != nil {
-		return nil, err
-	}
-	return &cfg, nil
+const CONFIG_TOML_PATH = "../../pkg/testfiles"
+const TEST_IP_1 = "192.0.2.11"
+
+type MockSSHUtilsImpl struct {
+	getSSHConfigFunc                                func() *SSHConfig
+	setSSHConfigFunc                                func(sshConfig *SSHConfig)
+	getClientConfigFunc                             func() (*ssh.ClientConfig, error)
+	getConnectionFunc                               func() (*ssh.Client, error)
+	connectAndExecuteCommandOnRemoteFunc            func(remoteCommands string, spinner bool) (string, error)
+	connectAndExecuteCommandOnRemoteSteamOutputFunc func(remoteCommands string) (string, error)
+	copyFileToRemoteFunc                            func(srcFilePath string, destFileName string, removeFile bool) error
+	copyFileFromRemoteFunc                          func(remoteFilePath string, outputFileName string) (string, error)
 }
 
-func TestDeleteNodeValidateError(t *testing.T) {
+func (msu *MockSSHUtilsImpl) getSSHConfig() *SSHConfig {
+	return msu.getSSHConfigFunc()
+}
+func (msu *MockSSHUtilsImpl) setSSHConfig(sshConfig *SSHConfig) {
+	return
+}
+func (msu *MockSSHUtilsImpl) getClientConfig() (*ssh.ClientConfig, error) {
+	return msu.getClientConfigFunc()
+}
+func (msu *MockSSHUtilsImpl) getConnection() (*ssh.Client, error) {
+	return msu.getConnectionFunc()
+}
+func (msu *MockSSHUtilsImpl) connectAndExecuteCommandOnRemote(remoteCommands string, spinner bool) (string, error) {
+	return msu.connectAndExecuteCommandOnRemoteFunc(remoteCommands, spinner)
+}
+func (msu *MockSSHUtilsImpl) connectAndExecuteCommandOnRemoteSteamOutput(remoteCommands string) (string, error) {
+	return msu.connectAndExecuteCommandOnRemoteSteamOutputFunc(remoteCommands)
+}
+func (msu *MockSSHUtilsImpl) copyFileToRemote(srcFilePath string, destFileName string, removeFile bool) error {
+	return msu.copyFileToRemoteFunc(srcFilePath, destFileName, removeFile)
+}
+func (msu *MockSSHUtilsImpl) copyFileFromRemote(remoteFilePath string, outputFileName string) (string, error) {
+	return msu.copyFileFromRemoteFunc(remoteFilePath, outputFileName)
+}
+
+func TestAddnodeValidateNotExistingInfra(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
 		automateIp: "10.2.1.67,ewewedw",
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
+		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
+			return nil, &SSHConfig{}, nil
+		},
+		getModeFromConfigFunc: func(path string) (string, error) {
+			return AWS_MODE, nil
+		},
+		pullAndUpdateConfigFunc: PullConfFunc,
+	}, CONFIG_TOML_PATH, &fileutils.MockFileSystemUtils{}, &MockSSHUtilsImpl{
+		connectAndExecuteCommandOnRemoteFunc: func(remoteCommands string, spinner bool) (string, error) {
+			return "", nil
+		},
+	})
+	err := nodeAdd.validate()
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "Unsupported deployment type. Please check "+CONFIG_TOML_PATH+"/config.toml")
+}
+
+func TestAddnodeValidateError(t *testing.T) {
+	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
+	flags := AddDeleteNodeHACmdFlags{
+		automateIp: "10.2.1.67,ewewedw",
+	}
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -36,15 +91,12 @@ func TestDeleteNodeValidateError(t *testing.T) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
+	err := nodeAdd.validate()
 	assert.Error(t, err)
-	assert.Contains(t, err.Error(), `IP address validation failed: 
-Unable to remove node. Automate instance count cannot be less than 1. Final count 0 not allowed.
-Automate Ip 10.2.1.67 is not present in existing list of ip addresses. Please use a different private ip.
-Automate Ip ewewedw is not present in existing list of ip addresses. Please use a different private ip.`)
+	assert.Contains(t, err.Error(), "IP address validation failed: \nIncorrect Automate IP address format for ip ewewedw")
 }
 
-func TestDeleteNodeValidateErrorMultiple(t *testing.T) {
+func TestAddnodeValidateErrorMultiple(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
 		automateIp:   "10.2.1.67,ewewedw",
@@ -52,7 +104,7 @@ func TestDeleteNodeValidateErrorMultiple(t *testing.T) {
 		postgresqlIp: "10.2.1.657,ewewedw",
 		opensearchIp: "10.2.1.61,ewewedw",
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -65,57 +117,48 @@ func TestDeleteNodeValidateErrorMultiple(t *testing.T) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
+	err := nodeAdd.validate()
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), `IP address validation failed: 
-Unable to remove node. Automate instance count cannot be less than 1. Final count 0 not allowed.
-Automate Ip 10.2.1.67 is not present in existing list of ip addresses. Please use a different private ip.
-Automate Ip ewewedw is not present in existing list of ip addresses. Please use a different private ip.
-Unable to remove node. Chef Server instance count cannot be less than 1. Final count -1 not allowed.
-Chef-Server Ip 10.2.1.637 is not present in existing list of ip addresses. Please use a different private ip.
-Chef-Server Ip ewewedw is not present in existing list of ip addresses. Please use a different private ip.
-Unable to remove node. OpenSearch instance count cannot be less than 3. Final count 2 not allowed.
-OpenSearch Ip 10.2.1.61 is not present in existing list of ip addresses. Please use a different private ip.
-OpenSearch Ip ewewedw is not present in existing list of ip addresses. Please use a different private ip.
-Unable to remove node. Postgresql instance count cannot be less than 3. Final count 1 not allowed.
-Postgresql Ip 10.2.1.657 is not present in existing list of ip addresses. Please use a different private ip.
-Postgresql Ip ewewedw is not present in existing list of ip addresses. Please use a different private ip.`)
+Incorrect Automate IP address format for ip ewewedw
+Incorrect Chef-Server IP address format for ip 10.2.1.637
+Incorrect Chef-Server IP address format for ip ewewedw
+Incorrect OpenSearch IP address format for ip ewewedw
+Incorrect Postgresql IP address format for ip 10.2.1.657
+Incorrect Postgresql IP address format for ip ewewedw`)
 }
 
-func TestDeleteNodeModifyAutomate(t *testing.T) {
+func TestAddnodeReadfileError(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
-		automateIp: "192.0.2.0",
+		automateIp: TEST_IP_1,
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
 		getModeFromConfigFunc: func(path string) (string, error) {
 			return EXISTING_INFRA_MODE, nil
 		},
-		pullAndUpdateConfigFunc: PullConfFunc,
+		pullAndUpdateConfigFunc: func(sshUtil *SSHUtil, exceptionIps []string) (*ExistingInfraConfigToml, error) {
+			return nil, errors.New("random")
+		},
 	}, CONFIG_TOML_PATH, &fileutils.MockFileSystemUtils{}, &MockSSHUtilsImpl{
 		connectAndExecuteCommandOnRemoteFunc: func(remoteCommands string, spinner bool) (string, error) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
-	assert.NoError(t, err)
-	err = nodedelete.modifyConfig()
-	assert.NoError(t, err)
-	assert.Equal(t, flags.automateIp, nodedelete.(*DeleteNodeImpl).automateIpList[0])
-	assert.Equal(t, "1", nodedelete.(*DeleteNodeImpl).config.Automate.Config.InstanceCount)
-	assert.Equal(t, 1, len(nodedelete.(*DeleteNodeImpl).config.Automate.Config.CertsByIP))
-	assert.Equal(t, 1, len(nodedelete.(*DeleteNodeImpl).config.ExistingInfra.Config.AutomatePrivateIps))
+	err := nodeAdd.validate()
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "random")
 }
 
-func TestRemovenodeValidateTypeAwsOrSelfManaged(t *testing.T) {
+func TestAddnodeValidateTypeAwsOrSelfManaged(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
 		postgresqlIp: TEST_IP_1,
 	}
-	nodeAdd := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -137,16 +180,16 @@ func TestRemovenodeValidateTypeAwsOrSelfManaged(t *testing.T) {
 	})
 	err := nodeAdd.validate()
 	assert.Error(t, err)
-	assert.Contains(t, err.Error(), fmt.Sprintf(TYPE_ERROR, "remove"))
+	assert.Contains(t, err.Error(), fmt.Sprintf(TYPE_ERROR, "add"))
 }
 
-func TestRemovenodeValidateTypeAwsOrSelfManaged2(t *testing.T) {
+func TestAddnodeValidateTypeAwsOrSelfManaged2(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
 		opensearchIp: TEST_IP_1,
 		automateIp:   TEST_IP_2,
 	}
-	nodeAdd := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -168,15 +211,15 @@ func TestRemovenodeValidateTypeAwsOrSelfManaged2(t *testing.T) {
 	})
 	err := nodeAdd.validate()
 	assert.Error(t, err)
-	assert.Contains(t, err.Error(), fmt.Sprintf(TYPE_ERROR, "remove"))
+	assert.Contains(t, err.Error(), fmt.Sprintf(TYPE_ERROR, "add"))
 }
 
-func TestDeleteNodeModifyInfra(t *testing.T) {
+func TestAddnodeModifyAutomate(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
-		chefServerIp: "192.0.2.2",
+		automateIp: TEST_IP_1,
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -189,24 +232,22 @@ func TestDeleteNodeModifyInfra(t *testing.T) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "Unable to remove node. Chef Server instance count cannot be less than 1. Final count 0 not allowed.")
-	// even though validation will fail still we check if modify config is working as expected or not
-	err = nodedelete.modifyConfig()
+	err := nodeAdd.validate()
 	assert.NoError(t, err)
-	assert.Equal(t, flags.chefServerIp, nodedelete.(*DeleteNodeImpl).chefServerIpList[0])
-	assert.Equal(t, "0", nodedelete.(*DeleteNodeImpl).config.ChefServer.Config.InstanceCount)
-	assert.Equal(t, 0, len(nodedelete.(*DeleteNodeImpl).config.ChefServer.Config.CertsByIP))
-	assert.Equal(t, 0, len(nodedelete.(*DeleteNodeImpl).config.ExistingInfra.Config.ChefServerPrivateIps))
+	err = nodeAdd.modifyConfig()
+	assert.NoError(t, err)
+	assert.Equal(t, flags.automateIp, nodeAdd.(*AddNodeOnPremImpl).automateIpList[0])
+	assert.Equal(t, "3", nodeAdd.(*AddNodeOnPremImpl).config.Automate.Config.InstanceCount)
+	assert.Equal(t, 3, len(nodeAdd.(*AddNodeOnPremImpl).config.Automate.Config.CertsByIP))
+	assert.Equal(t, 3, len(nodeAdd.(*AddNodeOnPremImpl).config.ExistingInfra.Config.AutomatePrivateIps))
 }
 
-func TestDeletenodeModifyOpensearch(t *testing.T) {
+func TestAddnodeModifyInfra(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
-		opensearchIp: "192.0.2.6",
+		chefServerIp: TEST_IP_1,
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -219,22 +260,22 @@ func TestDeletenodeModifyOpensearch(t *testing.T) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
+	err := nodeAdd.validate()
 	assert.NoError(t, err)
-	err = nodedelete.modifyConfig()
+	err = nodeAdd.modifyConfig()
 	assert.NoError(t, err)
-	assert.Equal(t, flags.opensearchIp, nodedelete.(*DeleteNodeImpl).opensearchIpList[0])
-	assert.Equal(t, "3", nodedelete.(*DeleteNodeImpl).config.Opensearch.Config.InstanceCount)
-	assert.Equal(t, 3, len(nodedelete.(*DeleteNodeImpl).config.Opensearch.Config.CertsByIP))
-	assert.Equal(t, 3, len(nodedelete.(*DeleteNodeImpl).config.ExistingInfra.Config.OpensearchPrivateIps))
+	assert.Equal(t, flags.chefServerIp, nodeAdd.(*AddNodeOnPremImpl).chefServerIpList[0])
+	assert.Equal(t, "2", nodeAdd.(*AddNodeOnPremImpl).config.ChefServer.Config.InstanceCount)
+	assert.Equal(t, 2, len(nodeAdd.(*AddNodeOnPremImpl).config.ChefServer.Config.CertsByIP))
+	assert.Equal(t, 2, len(nodeAdd.(*AddNodeOnPremImpl).config.ExistingInfra.Config.ChefServerPrivateIps))
 }
 
-func TestDeletenodeModifyPostgresql(t *testing.T) {
+func TestAddnodeModifyPostgresql(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
 	flags := AddDeleteNodeHACmdFlags{
-		postgresqlIp: "192.0.2.9",
+		postgresqlIp: TEST_IP_1,
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -247,24 +288,50 @@ func TestDeletenodeModifyPostgresql(t *testing.T) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "Unable to remove node. Postgresql instance count cannot be less than 3. Final count 2 not allowed.")
-	// even though validation will fail still we check if modify config is working as expected or not
-	err = nodedelete.modifyConfig()
+	err := nodeAdd.validate()
 	assert.NoError(t, err)
-	assert.Equal(t, flags.postgresqlIp, nodedelete.(*DeleteNodeImpl).postgresqlIp[0])
-	assert.Equal(t, "2", nodedelete.(*DeleteNodeImpl).config.Postgresql.Config.InstanceCount)
-	assert.Equal(t, 2, len(nodedelete.(*DeleteNodeImpl).config.Postgresql.Config.CertsByIP))
-	assert.Equal(t, 2, len(nodedelete.(*DeleteNodeImpl).config.ExistingInfra.Config.PostgresqlPrivateIps))
+	err = nodeAdd.modifyConfig()
+	assert.NoError(t, err)
+	assert.Equal(t, flags.postgresqlIp, nodeAdd.(*AddNodeOnPremImpl).postgresqlIp[0])
+	assert.Equal(t, "4", nodeAdd.(*AddNodeOnPremImpl).config.Postgresql.Config.InstanceCount)
+	assert.Equal(t, 4, len(nodeAdd.(*AddNodeOnPremImpl).config.Postgresql.Config.CertsByIP))
+	assert.Equal(t, 4, len(nodeAdd.(*AddNodeOnPremImpl).config.ExistingInfra.Config.PostgresqlPrivateIps))
 }
 
-func TestDeleteNodePrompt(t *testing.T) {
+func TestAddnodeModifyOpensearch(t *testing.T) {
+	w := majorupgrade_utils.NewCustomWriterWithInputs("x")
+	flags := AddDeleteNodeHACmdFlags{
+		opensearchIp: TEST_IP_1,
+	}
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
+		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
+			return nil, &SSHConfig{}, nil
+		},
+		getModeFromConfigFunc: func(path string) (string, error) {
+			return EXISTING_INFRA_MODE, nil
+		},
+		pullAndUpdateConfigFunc: PullConfFunc,
+	}, CONFIG_TOML_PATH, &fileutils.MockFileSystemUtils{}, &MockSSHUtilsImpl{
+		connectAndExecuteCommandOnRemoteFunc: func(remoteCommands string, spinner bool) (string, error) {
+			return "", nil
+		},
+	})
+	err := nodeAdd.validate()
+	assert.NoError(t, err)
+	err = nodeAdd.modifyConfig()
+	assert.NoError(t, err)
+	assert.Equal(t, flags.opensearchIp, nodeAdd.(*AddNodeOnPremImpl).opensearchIpList[0])
+	assert.Equal(t, "5", nodeAdd.(*AddNodeOnPremImpl).config.Opensearch.Config.InstanceCount)
+	assert.Equal(t, 5, len(nodeAdd.(*AddNodeOnPremImpl).config.Opensearch.Config.CertsByIP))
+	assert.Equal(t, 5, len(nodeAdd.(*AddNodeOnPremImpl).config.ExistingInfra.Config.OpensearchPrivateIps))
+}
+
+func TestAddnodePrompt(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("y")
 	flags := AddDeleteNodeHACmdFlags{
-		automateIp: "192.0.2.0",
+		automateIp: TEST_IP_1,
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -277,12 +344,12 @@ func TestDeleteNodePrompt(t *testing.T) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
+	err := nodeAdd.validate()
 	assert.NoError(t, err)
-	err = nodedelete.modifyConfig()
+	err = nodeAdd.modifyConfig()
 	assert.NoError(t, err)
-	assert.Equal(t, flags.automateIp, nodedelete.(*DeleteNodeImpl).automateIpList[0])
-	res, err := nodedelete.promptUserConfirmation()
+	assert.Equal(t, flags.automateIp, nodeAdd.(*AddNodeOnPremImpl).automateIpList[0])
+	res, err := nodeAdd.promptUserConfirmation()
 	assert.Equal(t, true, res)
 	assert.NoError(t, err)
 	assert.Contains(t, w.Output(), `Existing nodes:
@@ -292,19 +359,19 @@ Chef-Server => 192.0.2.2
 OpenSearch => 192.0.2.3, 192.0.2.4, 192.0.2.5, 192.0.2.6
 Postgresql => 192.0.2.7, 192.0.2.8, 192.0.2.9
 
-Nodes to be deleted:
+New nodes to be added:
 ================================================
-Automate => 192.0.2.0
-This will delete the above nodes from your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
+Automate => 192.0.2.11
+This will add the new nodes to your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
 }
 
-func TestDeleteNodeDeployWithNewOSNode(t *testing.T) {
+func TestAddnodeDeployWithNewOSNode(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("y")
 	flags := AddDeleteNodeHACmdFlags{
-		opensearchIp: "192.0.2.3",
+		opensearchIp: TEST_IP_1,
 	}
 	var filewritten, deployed bool
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -329,12 +396,12 @@ func TestDeleteNodeDeployWithNewOSNode(t *testing.T) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
+	err := nodeAdd.validate()
 	assert.NoError(t, err)
-	err = nodedelete.modifyConfig()
+	err = nodeAdd.modifyConfig()
 	assert.NoError(t, err)
-	assert.Equal(t, flags.opensearchIp, nodedelete.(*DeleteNodeImpl).opensearchIpList[0])
-	res, err := nodedelete.promptUserConfirmation()
+	assert.Equal(t, flags.opensearchIp, nodeAdd.(*AddNodeOnPremImpl).opensearchIpList[0])
+	res, err := nodeAdd.promptUserConfirmation()
 	assert.Equal(t, true, res)
 	assert.NoError(t, err)
 	assert.Contains(t, w.Output(), `Existing nodes:
@@ -344,22 +411,22 @@ Chef-Server => 192.0.2.2
 OpenSearch => 192.0.2.3, 192.0.2.4, 192.0.2.5, 192.0.2.6
 Postgresql => 192.0.2.7, 192.0.2.8, 192.0.2.9
 
-Nodes to be deleted:
+New nodes to be added:
 ================================================
-OpenSearch => 192.0.2.3
-This will delete the above nodes from your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
-	err = nodedelete.runDeploy()
+OpenSearch => 192.0.2.11
+This will add the new nodes to your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
+	err = nodeAdd.runDeploy()
 	assert.NoError(t, err)
 	assert.Equal(t, true, filewritten)
 	assert.Equal(t, true, deployed)
 }
 
-func TestDeleteNodeDeployWithNewOSMinCountError(t *testing.T) {
+func TestAddnodeDeployWithNewOSNodeGenconfigError(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("y")
 	flags := AddDeleteNodeHACmdFlags{
-		opensearchIp: "192.0.2.3,192.0.2.5",
+		opensearchIp: TEST_IP_1,
 	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -367,7 +434,7 @@ func TestDeleteNodeDeployWithNewOSMinCountError(t *testing.T) {
 			return nil
 		},
 		genConfigfunc: func(path string) error {
-			return nil
+			return errors.New("random")
 		},
 		getModeFromConfigFunc: func(path string) (string, error) {
 			return EXISTING_INFRA_MODE, nil
@@ -375,53 +442,19 @@ func TestDeleteNodeDeployWithNewOSMinCountError(t *testing.T) {
 		pullAndUpdateConfigFunc: PullConfFunc,
 	}, CONFIG_TOML_PATH, &fileutils.MockFileSystemUtils{
 		WriteToFileFunc: func(filepath string, data []byte) error {
-			return errors.New("random")
+			return nil
 		},
 	}, &MockSSHUtilsImpl{
 		connectAndExecuteCommandOnRemoteFunc: func(remoteCommands string, spinner bool) (string, error) {
 			return "", nil
 		},
 	})
-	err := nodedelete.validate()
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), `IP address validation failed: 
-Unable to remove node. OpenSearch instance count cannot be less than 3. Final count 2 not allowed.`)
-}
-
-func TestDeleteNodeDeployWithNewOSNodeError(t *testing.T) {
-	w := majorupgrade_utils.NewCustomWriterWithInputs("y")
-	flags := AddDeleteNodeHACmdFlags{
-		opensearchIp: "192.0.2.3",
-	}
-	nodedelete := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
-		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
-			return nil, &SSHConfig{}, nil
-		},
-		executeAutomateClusterCtlCommandAsyncfunc: func(command string, args []string, helpDocs string) error {
-			return nil
-		},
-		genConfigfunc: func(path string) error {
-			return nil
-		},
-		getModeFromConfigFunc: func(path string) (string, error) {
-			return EXISTING_INFRA_MODE, nil
-		},
-		pullAndUpdateConfigFunc: PullConfFunc,
-	}, CONFIG_TOML_PATH, &fileutils.MockFileSystemUtils{
-		WriteToFileFunc: func(filepath string, data []byte) error {
-			return errors.New("random")
-		},
-	}, &MockSSHUtilsImpl{
-		connectAndExecuteCommandOnRemoteFunc: func(remoteCommands string, spinner bool) (string, error) {
-			return "", nil
-		},
-	})
-	err := nodedelete.validate()
+	err := nodeAdd.validate()
 	assert.NoError(t, err)
-	err = nodedelete.modifyConfig()
+	err = nodeAdd.modifyConfig()
 	assert.NoError(t, err)
-	assert.Equal(t, flags.opensearchIp, nodedelete.(*DeleteNodeImpl).opensearchIpList[0])
-	res, err := nodedelete.promptUserConfirmation()
+	assert.Equal(t, flags.opensearchIp, nodeAdd.(*AddNodeOnPremImpl).opensearchIpList[0])
+	res, err := nodeAdd.promptUserConfirmation()
 	assert.Equal(t, true, res)
 	assert.NoError(t, err)
 	assert.Contains(t, w.Output(), `Existing nodes:
@@ -431,22 +464,22 @@ Chef-Server => 192.0.2.2
 OpenSearch => 192.0.2.3, 192.0.2.4, 192.0.2.5, 192.0.2.6
 Postgresql => 192.0.2.7, 192.0.2.8, 192.0.2.9
 
-Nodes to be deleted:
+New nodes to be added:
 ================================================
-OpenSearch => 192.0.2.3
-This will delete the above nodes from your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
-	err = nodedelete.runDeploy()
+OpenSearch => 192.0.2.11
+This will add the new nodes to your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
+	err = nodeAdd.runDeploy()
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "random")
 }
 
-func TestRemovenodeExecuteWithNewOSNodeNoCertsByIP(t *testing.T) {
+func TestAddnodeExecuteWithNewOSNodeNoCertByIP(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("y")
 	flags := AddDeleteNodeHACmdFlags{
-		opensearchIp: "192.0.2.6",
+		opensearchIp: TEST_IP_1,
 	}
 	var filewritten, deployed bool
-	nodeAdd := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -499,21 +532,21 @@ Chef-Server => 192.0.2.2
 OpenSearch => 192.0.2.3, 192.0.2.4, 192.0.2.5, 192.0.2.6
 Postgresql => 192.0.2.7, 192.0.2.8, 192.0.2.9
 
-Nodes to be deleted:
+New nodes to be added:
 ================================================
-OpenSearch => 192.0.2.6
-This will delete the above nodes from your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
+OpenSearch => 192.0.2.11
+This will add the new nodes to your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
 	assert.Equal(t, true, filewritten)
 	assert.Equal(t, true, deployed)
 }
 
-func TestRemovenodeExecuteWithNewOSNode(t *testing.T) {
+func TestAddnodeExecuteWithNewOSNode(t *testing.T) {
 	w := majorupgrade_utils.NewCustomWriterWithInputs("y")
 	flags := AddDeleteNodeHACmdFlags{
-		opensearchIp: "192.0.2.6",
+		opensearchIp: TEST_IP_1,
 	}
 	var filewritten, deployed bool
-	nodeAdd := NewDeleteNode(w.CliWriter, flags, &MockNodeUtilsImpl{
+	nodeAdd := NewAddNodeOnPrem(w.CliWriter, flags, &MockNodeUtilsImpl{
 		getHaInfraDetailsfunc: func() (*AutomteHAInfraDetails, *SSHConfig, error) {
 			return nil, &SSHConfig{}, nil
 		},
@@ -556,10 +589,10 @@ Chef-Server => 192.0.2.2
 OpenSearch => 192.0.2.3, 192.0.2.4, 192.0.2.5, 192.0.2.6
 Postgresql => 192.0.2.7, 192.0.2.8, 192.0.2.9
 
-Nodes to be deleted:
+New nodes to be added:
 ================================================
-OpenSearch => 192.0.2.6
-This will delete the above nodes from your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
+OpenSearch => 192.0.2.11
+This will add the new nodes to your existing setup. It might take a while. Are you sure you want to continue? (y/n)`)
 	assert.Equal(t, true, filewritten)
 	assert.Equal(t, true, deployed)
 }
